Added support for non-UTF8 data
If you are going to use this with the Magento EE Full Page Cache the binary data that results, caused by automatic gzipping, causes Mongo to reject the insertion, thus removing most of, if not all of, the benefit of using the FPC.  This change automatically converts non-UTF8-based content to a MongoBinData object on save and back to a string on load.  It does this by doing a preg_match('!!u') to see if it contains non-UTF8 data.  If so it will convert the string to MongoBinData.  When content is loaded it will check the data type of the requested object and return the binary representation.